[FLINK-21401] Consolidate different JobGraphTestUtils into one class
