--- flink-runtime/src/test/java/org/apache/flink/runtime/jobgraph/utils/JobGraphTestUtils.java
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.runtime.jobgraph.utils;
-
-import org.apache.flink.runtime.jobgraph.JobGraph;
-import org.apache.flink.runtime.jobgraph.JobVertex;
-import org.apache.flink.runtime.testtasks.NoOpInvokable;
-
-/** Test utilities for the {@link JobGraph}. */
-public enum JobGraphTestUtils {
-    ;
-
-    public static JobGraph createSingleVertexJobGraph() {
-        final JobVertex jobVertex = new JobVertex("Test vertex");
-        jobVertex.setInvokableClass(NoOpInvokable.class);
-
-        return new JobGraph(jobVertex);
-    }
-}
